+++ pkg/providers/gtm/provider.go
@@ -77,6 +77,7 @@ func (f *FrameworkSubprovider) Resources() []func() resource.Resource {
 // DataSources returns the GTM data sources implemented using terraform-plugin-framework
 func (f *FrameworkSubprovider) DataSources() []func() datasource.DataSource {
 	return []func() datasource.DataSource{
+		NewGTMAsmapDataSource,
 		NewGTMDomainDataSource,
 		NewGTMDomainsDataSource,
 		NewGTMResourceDataSource,